termin/requirements.txt
requests
selenium
python-telegram-bot==13.7
bs4
ddddocr
python-dotenv
webdriver_manager